requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.7
pillow==10.0.0